fix: 스크리닝 만점을 13점 대신 10점으로 재조정 (조건 유지, 배점만 조절)
조건을 빼는 대신 각 항목 배점을 조절해 정확히 10점 만점으로 재조정:
- 추세 전환 신호(MA60, MA5>MA20, 골든크로스, RSI 과매도 반등, 바닥 다지기,
  거래량 감소 추세) — 각 1점 유지. 하락의 끝을 알리는 핵심 신호로 보고
  비중을 그대로 둠.
- 거래량 충분(유동성) — 1점 유지.
- PER·PBR·부채비율·유동비율·현금비율·이자보상배율 — 급락으로 왜곡되기
  쉬운 보조 지표라 각 0.5점으로 낮춤.
  (1+1+1+1+1+1) + 1 + (0.5×6) = 10.0
- BUY_THRESHOLD 6점으로 원복 (10점 만점 60% 비율 유지).

Supabase screening_results/us_screening_results의 score 컬럼을 INTEGER에서
NUMERIC(4,1)로 변경 — 0.5점 단위 점수를 저장하려면 필수 (프로덕션에 직접 적용 완료).

+++ trading/board_post.py
@@ -83,7 +83,7 @@ def screening_content(results: list, date_str: str) -> str:
 def buy_content(
     stock_name: str, stock_code: str,
     price: int, buy_amount: int, shares: int,
-    score: int, max_score: int,
+    score: float, max_score: int,
     log_lines: list,
     sell_opinion: str,
     ratios: dict,
